requirements.txt
python-telegram-bot[job-queue]==21.6
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
alembic==1.13.2
pydantic-settings==2.4.0
fastapi==0.112.2
uvicorn[standard]==0.30.6
httpx==0.27.2
redis==5.0.8
python-dotenv==1.0.1
pytest==8.3.2
pytest-asyncio==0.23.8

# Phase II (uncomment when ready)
# celery==5.4.0
# stripe==10.8.0
# chapa-python==1.0.0
